Refactor page view tracking and identification in their own services.
