+++ Website/src/rendering/src/components/Checkout/CheckoutDetails.tsx
@@ -12,12 +12,12 @@ import CheckoutSummary from './CheckoutSummary';
 import NoItemsInCartMessage from '../ShopCommon/NoItemsInCartMessage';
 import useOcCurrentCart from '../../hooks/useOcCurrentCart';
 import useOcAuth from '../../hooks/useOcAuth';
-import { getGuestEmail, identifyVisitor } from '../../services/CdpService';
+import { getGuestEmail } from '../../services/CdpService';
+import { identifyVisitor } from '../../services/IdentificationService';
 import { useAppDispatch } from '../../redux/store';
 import { updateUser } from '../../redux/ocUser';
 import { patchOrder } from '../../redux/ocCurrentCart';
 import { DeliveryTypes } from '../../models/ordercloud/DOrder';
-import { identifyVisitorInSend } from '../../services/SendService';
 
 const CheckoutDetailsSkeleton = (): JSX.Element => {
   const skeletonCount = 5;
@@ -65,7 +65,6 @@ const CheckoutDetails = (): JSX.Element => {
   }, [isAnonymous, setEmail]);
 
   const handleReviewOrderClick = () => {
-    identifyVisitorInSend(isAnonymous ? userEmail : order.FromUser.Email);
     identifyVisitor(isAnonymous ? userEmail : order.FromUser.Email);
     return router?.push('/shop/checkout/order-review');
   };
