+++ Website/src/rendering/src/components/Products/Shop.tsx
@@ -9,17 +9,18 @@ import { Provider } from 'react-redux';
 import reduxStore from '../../redux/store';
 import OcProvider from '../../redux/ocProvider';
 import { DiscoverService } from '../../services/DiscoverService';
-import { logViewEvent } from '../../services/CdpService';
+import { trackViewEvent } from '../../services/TrackingService';
 
 DiscoverService();
 
 export const ShopLayout = (props: PropsWithChildren<unknown>): JSX.Element => {
   useEffect(() => {
-    // Log a CDP page view on route change
+    trackViewEvent();
+    // Log a page view on route change
     const pushState = history.pushState;
     history.pushState = (...rest) => {
       pushState.apply(history, rest);
-      logViewEvent();
+      trackViewEvent();
     };
   }, []);
