+++ Website/src/rendering/src/services/TrackingService.ts
@@ -1,0 +1,8 @@
+import { RouteData } from '@sitecore-jss/sitecore-jss-nextjs';
+import { logViewEvent as trackViewEventInCdp } from './CdpService';
+import { trackViewEvent as trackViewEventInSend } from './SendService';
+
+export function trackViewEvent(route?: RouteData): Promise<unknown> {
+  trackViewEventInSend();
+  return trackViewEventInCdp(route);
+}
